requirements.txt

Django==3.2.10
djangorestframework==3.12.4
requests==2.26.0
django-cors-headers==3.7.0
gunicorn==20.1.0
whitenoise==5.3.0
psycopg2-binary==2.9.1
